requirements.txt
requests
beautifulsoup4
tqdm
gspread
oauth2client
python-dotenv
openai